+++ tests/TestCase.php
@@ -105,7 +105,7 @@ abstract class TestCase extends Framework\TestCase
                 return $faker->uuid;
             case 'dueTime':
             case 'expiredTime':
-            case 'expireTime': // TODO inconsistent name
+            case 'expirationTime':
             case 'periodStartTime':
             case 'renewalTime':
             case 'periodEndTime':